Updated render method, removed unneeded component, merged iframe into sandbox
